% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell_solver.R
\name{adaptive_shell_solver}
\alias{adaptive_shell_solver}
\title{shell_solver with precision}
\usage{
adaptive_shell_solver(
  cmd,
  nparams,
  qoi_file,
  required_args = "precision",
  arg_combine_fn = function(x, precision, ...) paste(c(x, precision), collapse = " "),
  ...
)
}
\arguments{
\item{cmd}{command to run the solver executable.}

\item{nparams}{numeric, number of parameters, must be provided.}

\item{qoi_file}{name of file containing the computed quantity of interest,
may be absolute or relative to \code{wd}, cannot be \code{NULL}.}

\item{required_args}{character, mandatory args for \code{\link[=run]{run()}}.}

\item{arg_combine_fn}{function producing appropriate numeric scalar
from point and precision for solver command line}

\item{...}{other args passed to \code{\link[=shell_solver]{shell_solver()}}.}
}
\description{
A shell_solver that can compute the qoi with expected precision. \code{precision}
becomes mandatory argument of \code{\link[=run]{run()}} method.
}
\examples{
rscript_path <- file.path(R.home(), "bin", "Rscript")
solver_path <-
  file.path(find.package("solvergater"), "exec", "fake_adaptive.R")
nparams <- 2
nqoi <- 5
solver_cmd <- paste(rscript_path, solver_path, nparams, nqoi)
s <- adaptive_shell_solver(solver_cmd, nparams = nparams,
qoi_file = "output_qoi", jacobian_file = "output_jacobian", wd = tempdir())
run(s, c(20, 5), precision = 90)
}
