% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objective.R
\name{assert_arg_not_null}
\alias{assert_arg_not_null}
\title{Assert that an arg is not \code{NULL}}
\usage{
assert_arg_not_null(arg_value, arg_name)
}
\arguments{
\item{arg_value}{argument value}

\item{arg_name}{argument name}
}
\description{
Assert that an arg is not \code{NULL}
}
