% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_utils.R
\name{assert_point_not_null}
\alias{assert_point_not_null}
\title{Assert that point is not \code{NULL}}
\usage{
assert_point_not_null(x)
}
\arguments{
\item{x}{numeric vector representing point}
}
\description{
Assert that point is not \code{NULL}
}
