% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objective.R
\name{assert_precision_not_null}
\alias{assert_precision_not_null}
\title{Assert that precision is not \code{NULL}}
\usage{
assert_precision_not_null(precision)
}
\arguments{
\item{precision}{lenght 1 numeric vector representing precision}
}
\description{
Assert that precision is not \code{NULL}
}
