% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objective.R, R/extsolver_none.R,
%   R/extsolver_simple.R
\name{compute_objective}
\alias{compute_objective}
\alias{compute_objective.extsolver_none}
\alias{compute_objective.extsolver_simple}
\title{Compute the objective}
\usage{
compute_objective(solver, x, precision)

\method{compute_objective}{extsolver_none}(solver, x, precision = NULL)

\method{compute_objective}{extsolver_simple}(solver, x, precision)
}
\arguments{
\item{solver}{solver object}

\item{x}{numeric vector representing the point to compute objective at}

\item{precision}{length 1 numeric vector indicating desired solver accuracy (solver-dependent)}
}
\value{
List with components:
\itemize{
\item \code{value} objective value at \code{x}, default implementation returns \code{NA};
\item \code{gradient} gradient of the objective at \code{x}, default implementation returns \code{NA},
\code{NA} is also returned if the solver does not provide gradient info.
}
}
\description{
Computes the objective value and gradient using approprate type of external
solver.
}
\section{Methods (by class)}{
\itemize{
\item \code{extsolver_none}: Computes output using provided functions, ignores
\code{precision}.

\item \code{extsolver_simple}: Runs solver executable and reads values from output
file(s).
}}

\examples{
compute_objective(x = 10)
compute_objective(NULL, c(2, 4))
}
