% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extsolver_simple.R
\name{extsolver_simple}
\alias{extsolver_simple}
\title{Simple external solver gateway}
\usage{
extsolver_simple(
  cmd,
  value_file,
  gradient_file = NULL,
  value_read_fn = function(file) scan(file, quiet = TRUE),
  gradient_read_fn = if (!is.null(gradient_file)) value_read_fn
)
}
\arguments{
\item{cmd}{command to run the solver executable}

\item{value_file}{name of file containing the computed objective value, cannot
be \code{NULL}}

\item{gradient_file}{name of file containing the computed objective gradient,
\code{NULL} indicates that the solver does not provide gradient info}

\item{value_read_fn}{function reading the objective value from file}

\item{gradient_read_fn}{function reading the objective gradient from file}
}
\value{
An object of classes \code{extsolver_simple} and \code{extsolver}
}
\description{
Gateway to simple external solver that stores objective value and gradient
in separate files with simple format (real numbers separated with whitespaces).
}
\examples{
rscript_path <- file.path(R.home(), "bin", "Rscript")
solver_path <- file.path(find.package("solvergater"), "exec", "fake_simple.R")
solver_cmd <- paste(rscript_path, solver_path)
s <- extsolver_simple(solver_cmd, value_file = "output_value", gradient_file = "output_gradient")
old_wd <- getwd()
setwd(tempdir())
compute_objective(s, c(20, 5), 10)
setwd(old_wd)
}
