% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_solvers.R
\name{fake_adaptive_solver}
\alias{fake_adaptive_solver}
\title{Gateway for \code{fake_adaptive.R} mock solver}
\usage{
fake_adaptive_solver(nparams, nqoi, wd = tempdir(), ...)
}
\arguments{
\item{nparams}{numeric scalar, number of parameters.}

\item{nqoi}{numeric scalar, number of quantities of interest.}

\item{wd}{character, working directory.}

\item{...}{additional parameters passed to \code{\link[=shell_solver]{shell_solver()}}.}
}
\description{
A helper to create \code{solver} object encapsulating calls to \code{fake_adaptive.R}.
}
\examples{
s <- fake_adaptive_solver(3, 4)
run(s, c(-1, 5.2, 0), precision = 100)
}
