% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{make_functions}
\alias{make_functions}
\title{Make two functions out of objective}
\usage{
make_functions(objective, provides_gradient = TRUE)
}
\arguments{
\item{objective}{objective function, as created with \code{\link[=objective]{objective()}}.}

\item{provides_gradient}{logical, does \code{objective} prodive gradient?}
}
\value{
same as in \code{\link[=objective_functions]{objective_functions()}}.
}
\description{
Separates function created with \code{\link[=objective]{objective()}} into value function
and gradient function. No memoisation is done here.
}
\examples{
s <- fake_simple_solver(4, 5)
result <- run(s, c(10, 10, 10, 10), precision = 5.0, silent = TRUE)
observed_data <- result$qoi
x <- c(10.5, 9.44, 10.21, 8.14)
obj <- objective(s, observed_data, precision = 30.0, silent = TRUE)
solver_funs <- make_functions(obj)
solver_funs$value(x)
solver_funs$gradient(x)
}
