% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{nparams<-}
\alias{nparams<-}
\title{Set number of solver parameters}
\usage{
nparams(solver) <- value
}
\arguments{
\item{solver}{object of class \code{solver}}

\item{value}{numeric, can be \code{NULL}}
}
\description{
Set number of solver parameters
}
