% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell_solver.R
\name{output_file}
\alias{output_file}
\title{Get solver output file name}
\usage{
output_file(solver, type = c("value", "gradient"))
}
\arguments{
\item{solver}{object of class \code{shell_solver}}

\item{type}{character, file type}
}
\description{
This accessor function should be used instead of touching the components
of \code{shell_solver} object. It takes into account the working directory.
}
