% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extsolver_none.R, R/provides_gradient.R,
%   R/shell_solver.R
\name{provides_gradient.extsolver_none}
\alias{provides_gradient.extsolver_none}
\alias{provides_gradient}
\alias{provides_gradient.shell_solver}
\title{Check ability to provide objective gradient}
\usage{
\method{provides_gradient}{extsolver_none}(solver)

provides_gradient(solver)

\method{provides_gradient}{shell_solver}(solver)
}
\arguments{
\item{solver}{object of class \code{extsolver}}
}
\value{
logical, TRUE if the solver computes the objective gradient
}
\description{
Check ability to provide objective gradient
}
\section{Methods (by class)}{
\itemize{
\item \code{extsolver_none}: \code{TRUE} if gradient function has been provided

\item \code{shell_solver}: \code{TRUE} if gradient file has been provided
}}

\examples{
s <- extsolver_none()
provides_gradient(s)
}
