% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_utils.R
\name{read_matrix}
\alias{read_matrix}
\title{Read matrix from file}
\usage{
read_matrix(ncol)
}
\arguments{
\item{ncol}{integer, expected number of columns, passed to \code{\link[base:matrix]{base::matrix()}}.}
}
\value{
function with 1 argument: \code{file}
}
\description{
A function factory to read matrix with given number of columns from a file. The
matrix is read row-wise (with \code{byrow = TRUE}).
}
\examples{
m <- matrix(1:6, ncol = 3)
tf <- tempfile()
write(t(m), file = tf, ncolumns = 3)
rm3 <- read_matrix(ncol = 3)
rm3(tf)
}
