% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\alias{run.solver}
\alias{run.r_solver}
\alias{run.shell_solver}
\title{Run the solver}
\usage{
run(solver, x, precision, ...)

\method{run}{solver}(solver, x, precision, ...)

\method{run}{r_solver}(solver, x, precision = NULL, ...)

\method{run}{shell_solver}(solver, x, precision, ...)
}
\arguments{
\item{solver}{solver object}

\item{x}{numeric vector representing the point to compute objective at}

\item{precision}{length 1 numeric vector indicating desired solver accuracy (solver-dependent)}

\item{...}{additional arguments passed to other functions}
}
\value{
List with components:
\itemize{
\item \code{value} objective value at \code{x}, default implementation returns \code{NA};
\item \code{gradient} gradient of the objective at \code{x}, default implementation returns \code{NA},
\code{NA} is also returned if the solver does not provide gradient info.
}
}
\description{
Runs the solver executable and extracts the objective value and gradient.
}
\section{Methods (by class)}{
\itemize{
\item \code{solver}: Checks if number of solver parameters (when not
\code{NULL} equals length of given point) and delegates to default method.

\item \code{r_solver}: Computes output using provided functions, ignores
\code{precision}.

\item \code{shell_solver}: Runs solver executable and reads values from output
file(s). If solver process exits with non-zero status code, a warning is issued
and list of \code{NA}'s is returned.
}}

\examples{
run(x = 10)
run(NULL, c(2, 4))
}
