% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell_solver.R
\name{shell_solver}
\alias{shell_solver}
\title{Gateway to solver run from command line}
\usage{
shell_solver(
  cmd,
  qoi_file,
  jacobian_file = NULL,
  qoi_read_fn = function(file) scan(file, quiet = TRUE),
  jacobian_read_fn = if (!is.null(jacobian_file)) qoi_read_fn,
  arg_combine_fn = function(x, precision) paste(c(x, precision), collapse = " "),
  wd = NULL,
  ignore.stdout = TRUE,
  ignore.stderr = TRUE,
  nparams = NULL
)
}
\arguments{
\item{cmd}{command to run the solver executable.}

\item{qoi_file}{name of file containing the computed quantity of interest, can
be absolute or relative to \code{wd}, cannot be \code{NULL}.}

\item{jacobian_file}{name of file containing the computed Jacobian matrix
of quantity of interest,
can be absolute or relative to \code{wd}, \code{NULL} indicates that the solver does not
provide gradient info.}

\item{qoi_read_fn}{function reading the quantity of interest from file.}

\item{jacobian_read_fn}{function reading the Jacobian matrix from file.}

\item{arg_combine_fn}{function producing appropriate length-1 character vector
from point and precision for solver command line}

\item{wd}{character, working directory, i.e. directory where actual solver exec
has to be run, when \code{NULL} current working directory will be used}

\item{ignore.stdout}{logical, should solver STDOUT be ignored?}

\item{ignore.stderr}{logical, should solver STDERR be ignored?}

\item{nparams}{numeric, number of parameters, unspecified if \code{NULL}}
}
\value{
An object of classes \code{shell_solver} and \code{solver}
}
\description{
Gateway to simple external solver that is run from the command line
and stores quantity of interest and its jacobian (with respect to problem
parameters in output text files.
}
\examples{
rscript_path <- file.path(R.home(), "bin", "Rscript")
solver_path <- file.path(find.package("solvergater"), "exec", "fake_simple.R")
solver_cmd <- paste(rscript_path, solver_path)
s <- shell_solver(solver_cmd, qoi_file = "output_qoi",
jacobian_file = "output_jacobian", wd = tempdir())
run(s, c(20, 5), 10)
}
