% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvergater.R
\docType{package}
\name{solvergater}
\alias{solvergater}
\title{solvergater: External solver gateway in R}
\description{
The solvergater package provides the following categories of functions:
solver handlers, objective providers and helpers.
}
\section{Solver handlers}{

\itemize{
\item \code{\link[=r_solver]{r_solver()}}
\item \code{\link[=shell_solver]{shell_solver()}}
\item \code{\link[=fake_simple_solver]{fake_simple_solver()}}
\item \code{\link[=run]{run()}}
\item \code{\link[=nparams]{nparams()}}
\item \code{\link[='nparams<-']{'nparams<-'()}}
\item \code{\link[=provides_jacobian]{provides_jacobian()}}
\item \code{\link[=run_count]{run_count()}}
}
}

\section{Objective providers}{

\itemize{
\item \code{\link[=objective]{objective()}}
\item \code{\link[=objective_functions]{objective_functions()}}
\item \code{\link[=make_functions]{make_functions()}}
\item \code{\link[=differentiable]{differentiable()}}
\item \code{\link[=squared_error]{squared_error()}}
}
}

\section{Helpers}{

\itemize{
\item \code{\link[=read_matrix]{read_matrix()}}
\item \code{\link[=write_matrix]{write_matrix()}}
\item \code{\link[=output_file]{output_file()}}
}
}

