% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_utils.R
\name{write_matrix}
\alias{write_matrix}
\title{Write a matrix to a file}
\usage{
write_matrix(mat, file)
}
\arguments{
\item{mat}{matrix to be written}

\item{file}{file to write to}
}
\description{
For a matrix with N columns, \code{write_matrix()} and \code{read_matrix(N)} are thought as
mutual inverse.
}
\examples{
m <- matrix(1:6, ncol = 3)
tf <- tempfile()
write_matrix(m, tf)
rm3 <- read_matrix(3)
rm3(tf)
}
\seealso{
\code{\link[=read_matrix]{read_matrix()}}
}
