% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objective.R, R/r_solver.R,
%   R/shell_solver.R, R/solver.R
\name{compute_objective}
\alias{compute_objective}
\alias{compute_objective.r_solver}
\alias{compute_objective.shell_solver}
\alias{compute_objective.solver}
\title{Compute the objective}
\usage{
compute_objective(solver, x, precision, ...)

\method{compute_objective}{r_solver}(solver, x, precision = NULL, ...)

\method{compute_objective}{shell_solver}(solver, x, precision, ...)

\method{compute_objective}{solver}(solver, x, precision, ...)
}
\arguments{
\item{solver}{solver object}

\item{x}{numeric vector representing the point to compute objective at}

\item{precision}{length 1 numeric vector indicating desired solver accuracy (solver-dependent)}

\item{...}{additional arguments passed to other functions}
}
\value{
List with components:
\itemize{
\item \code{value} objective value at \code{x}, default implementation returns \code{NA};
\item \code{gradient} gradient of the objective at \code{x}, default implementation returns \code{NA},
\code{NA} is also returned if the solver does not provide gradient info.
}
}
\description{
Computes the objective value and gradient using approprate type of external
solver.
}
\section{Methods (by class)}{
\itemize{
\item \code{r_solver}: Computes output using provided functions, ignores
\code{precision}.

\item \code{shell_solver}: Runs solver executable and reads values from output
file(s). If solver process exits with non-zero status code, a warning is issued
and list of \code{NA}'s is returned.

\item \code{solver}: Checks if number of solver parameters (when not
\code{NULL} equals length of given point and delegates to default method).
}}

\examples{
compute_objective(x = 10)
compute_objective(NULL, c(2, 4))
}
