% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extsolver_simple.R
\name{extsolver_simple}
\alias{extsolver_simple}
\title{Simple external solver gateway}
\usage{
extsolver_simple(
  cmd,
  value_file = "output_value",
  gradient_file = "output_gradient"
)
}
\arguments{
\item{cmd}{command to run the solver executable (must point to something existing
and really executable)}

\item{value_file}{file containing the computed objective value}

\item{gradient_file}{file containing the computed objective gradient}
}
\value{
An object of classes \code{extsolver_simple} and \code{extsolver} with components
\itemize{
\item \code{cmd} command to run solver executable
\item \code{value_file} path to file containing objective value
\item \code{gradient_file} path to file containing objective gradient
}
}
\description{
Gateway to simple external solver that stores objective value and gradient
in separate files with simple format (real numbers separated with whitespaces).
}
\examples{
rscript_path <- file.path(R.home(), "bin", "Rscript")
solver_path <- file.path(find.package("solvergater"), "exec", "fake_simple.R")
s <- extsolver_simple(paste(rscript_path, solver_path))
old_wd <- getwd()
setwd(tempdir())
compute_objective(s, c(20, 5), 10)
setwd(old_wd)
}
