% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{objective}
\alias{objective}
\title{Solver objective}
\usage{
objective(solver, data, misfit_fn = lsq_misfit)
}
\arguments{
\item{solver}{object of class \code{solver}}

\item{data}{observed ('exact') data}

\item{misfit_fn}{function to compute misfit between \code{data} and result of simulation}
}
\value{
List with one or two components:
\itemize{
\item \code{value} objective value function,
\item \code{gradient} objective gradient function.
}
}
\description{
Function factory producing functions for use in optimization. These are
objective function and its gradient. They can be used, e.g., as parameters
\code{fn} and \code{gr} of \code{\link[stats:optim]{stats::optim()}}. Their parameters \code{...} are passed to
\code{\link[=run]{run()}}: note that some solvers can require some parameters, e.g.
\code{shell_solver} requires \code{precision}.
}
\examples{
s <- fake_simple_solver(4, 5)
observed_data <- run(s, c(10, 10, 10, 10), precision = 5.0)$qoi
x <- c(10.5, 9.44, 10.21, 8.14)
solver_obj <- objective(s, observed_data)
solver_obj$value(x, precision = 30.0)
solver_obj$gradient(x, precision = 30.0)
}
