% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{objective_functions}
\alias{objective_functions}
\title{Solver objective functions}
\usage{
objective_functions(solver, data, misfit_fn = lsq_misfit, ...)
}
\arguments{
\item{solver}{object of class \code{solver}.}

\item{data}{observed ('exact') data.}

\item{misfit_fn}{function to compute misfit between \code{data} and result of
simulation.}

\item{...}{additional args passed to \code{\link[=run]{run()}}, note that some solvers can
require some parameters, e.g. \code{shell_solver} requires \code{precision}.}
}
\value{
List with one or two components:
\itemize{
\item \code{value} objective value function,
\item \code{gradient} objective gradient function, missing if solver does not compute
Jacobian matrix.
}
}
\description{
Function factory producing functions for use in optimization. These are
objective function and its gradient. They can be used, e.g., as parameters
\code{fn} and \code{gr} of \code{\link[stats:optim]{stats::optim()}}.
}
\examples{
s <- fake_simple_solver(4, 5)
result <- run(s, c(10, 10, 10, 10), precision = 5.0, silent = TRUE)
observed_data <- result$qoi
x <- c(10.5, 9.44, 10.21, 8.14)
solver_funs <- objective_functions(s, observed_data, precision = 30.0,
silent = TRUE)
solver_funs$value(x)
solver_funs$gradient(x)
}
