% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provides_gradient.R, R/r_solver.R,
%   R/shell_solver.R
\name{provides_gradient}
\alias{provides_gradient}
\alias{provides_gradient.r_solver}
\alias{provides_gradient.shell_solver}
\title{Check ability to provide objective gradient}
\usage{
provides_gradient(solver)

\method{provides_gradient}{r_solver}(solver)

\method{provides_gradient}{shell_solver}(solver)
}
\arguments{
\item{solver}{object of class \code{solver}}
}
\value{
logical, TRUE if the solver computes the objective gradient
}
\description{
Check ability to provide objective gradient
}
\section{Methods (by class)}{
\itemize{
\item \code{r_solver}: \code{TRUE} if gradient function has been provided

\item \code{shell_solver}: \code{TRUE} if gradient file has been provided
}}

\examples{
s <- r_solver()
provides_gradient(s)
}
