% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{provides_gradient}
\alias{provides_gradient}
\title{Does this solver provide derivative info?}
\usage{
provides_gradient(solver)
}
\arguments{
\item{solver}{object of class \code{solver}}
}
\value{
logical, TRUE if the solver computes the objective gradient
}
\description{
Does this solver provide derivative info?
}
\examples{
s <- r_solver()
provides_gradient(s)
}
