% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_solver.R
\name{r_solver}
\alias{r_solver}
\title{Gateway for solver written in R}
\usage{
r_solver(
  objective = function(x) 0,
  gradient = function(x) rep(0, length(x)),
  nparams = NULL
)
}
\arguments{
\item{objective}{function to act as the objective (should return length-one vector)}

\item{gradient}{function to act as the objective gradient (should return vector
of the same lentgh as its parameter and should be the real gradient function
for \code{objective})}

\item{nparams}{numeric, number of parameters, unspecified if \code{NULL}}
}
\value{
An object of classes \code{r_solver} and \code{solver}
}
\description{
(Mock) gateway to solver that does not perform any external calls
and computes the objective and its gradient using provided functions.
Provided mainly for testing purposes.
}
\examples{
s <- r_solver(objective = function(x) sum(x^2), gradient = function(x) 2 * x)
compute_objective(s, c(10, 1.5))
}
