% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\alias{run.solver}
\alias{run.r_solver}
\alias{run.shell_solver}
\title{Run the solver}
\usage{
run(solver, x, ...)

\method{run}{solver}(solver, x, ...)

\method{run}{r_solver}(solver, x, ...)

\method{run}{shell_solver}(solver, x, precision, ...)
}
\arguments{
\item{solver}{solver object}

\item{x}{numeric vector representing the point to compute objective at}

\item{...}{additional arguments passed to other functions}

\item{precision}{positive numeric scalar, expected solver accuracy (if applicable)}
}
\value{
List with components:
\itemize{
\item \code{qoi} quantity of interest (QOI) at \code{x}, default implementation returns \code{NA};
\item \code{jacobian} QOI Jacobian matrix at \code{x}, default implementation returns \code{NA},
\code{NA} is also returned if the solver does not provide Jacobian info.
}
}
\description{
Runs the solver executable and extracts the quantity of interest and its
Jacobian matrix (if available).
}
\section{Methods (by class)}{
\itemize{
\item \code{solver}: Checks if number of solver parameters (when not
\code{NULL} equals length of given point) and delegates to default method.

\item \code{r_solver}: Computes output using provided functions.

\item \code{shell_solver}: Runs solver executable and reads values from output
file(s). If solver process exits with non-zero status code, a warning is issued
and list of \code{NA}'s is returned.
}}

\examples{
run(x = 10)
run(NULL, c(2, 4))
}
