% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{solver_attributes}
\alias{solver_attributes}
\alias{provides_gradient}
\alias{nparams}
\alias{nparams<-}
\title{Solver attribute accessors}
\usage{
provides_gradient(solver)

nparams(solver)

nparams(solver) <- value
}
\arguments{
\item{solver}{object of class \code{solver}}

\item{value}{numeric scalar, can be \code{NULL}}
}
\description{
Functions to get and set values of the most useful attributes of \code{solver} objects.
}
\examples{
s <- r_solver(nparams = 10)
provides_gradient(s)
nparams(s)
}
